+++ src/Psr7/Stream/Factory.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Utopia\Psr7\Stream;
 
-use InvalidArgumentException;
 use Psr\Http\Message\StreamFactoryInterface;
 use Psr\Http\Message\StreamInterface;
 use RuntimeException;
@@ -19,29 +18,17 @@ final class Factory implements StreamFactoryInterface
 
     public function createStreamFromFile(string $filename, string $mode = 'r'): StreamInterface
     {
-        $content = file_get_contents($filename);
+        $resource = @fopen($filename, $mode);
 
-        if ($content === false) {
-            throw new RuntimeException('Unable to read stream file.');
+        if (!\is_resource($resource)) {
+            throw new RuntimeException('Unable to open stream file.');
         }
 
-        unset($mode);
-
-        return new Stream($content);
+        return Stream::fromResource($resource);
     }
 
     public function createStreamFromResource($resource): StreamInterface
     {
-        if (!\is_resource($resource)) {
-            throw new InvalidArgumentException('Expected a valid stream resource.');
-        }
-
-        $content = stream_get_contents($resource);
-
-        if ($content === false) {
-            throw new RuntimeException('Unable to read stream resource.');
-        }
-
-        return new Stream($content);
+        return Stream::fromResource($resource);
     }
 }